+++ src/main/kotlin/omok/controller/OmokGame.kt
@@ -3,7 +3,7 @@ package omok.controller
 import omok.domain.Board
 import omok.domain.Stone
 import omok.domain.Turn
-import omok.model.RenjuRuleAdapter
+import omok.domain.RenjuRuleAdapter
 import omok.view.InputView
 import omok.view.OutputView
